+++ src/Main.js
@@ -2,49 +2,12 @@ import React, { useEffect, forwardRef } from 'react';
 import { useSet } from './hooks';
 // import SCHEMA from './json/basic.json';
 import FRWrapper from './FRWrapper';
-import { widgets } from './widgets/antd';
+import { widgets as defaultWidgets } from './widgets/antd';
 import { mapping } from './mapping';
 import './atom.css';
 import './Main.css';
 import 'antd/dist/antd.css';
 import { oldSchemaToNew } from './utils';
-
-// const SCHEMA = {
-//   schema: {
-//     type: 'object',
-//     properties: {
-//       obj1: {
-//         title: '对象',
-//         type: 'object',
-//       },
-//       obj2: {
-//         title: '对象',
-//         type: 'object',
-//         properties: {
-//           obj3: {
-//             title: '对象',
-//             type: 'object',
-//             properties: {
-//               col: {
-//                 title: '颜色选择',
-//                 type: 'string',
-//                 format: 'color',
-//               },
-//               img: {
-//                 title: '图片展示',
-//                 type: 'string',
-//                 format: 'image',
-//                 'ui:options': {},
-//               },
-//             },
-//           },
-//         },
-//       },
-//     },
-//   },
-//   uiSchema: {},
-//   formData: {},
-// };
 
 const DEFAULT_SCHEMA = {
   schema: {
@@ -56,15 +19,20 @@ const DEFAULT_SCHEMA = {
 };
 
 // TODO: formData 不存在的时候会报错：can't find # of undefined
-
 function App(
-  { defaultValue, templates, submit, transformer, extraButtons },
+  {
+    defaultValue,
+    templates,
+    submit,
+    transformer,
+    extraButtons,
+    settings,
+    commonSettings,
+    globalSettings,
+    widgets = {},
+  },
   ref,
 ) {
-  const initGlobal = {
-    displayType: 'row',
-  };
-
   let transformFrom = a => a;
   let transformTo = a => a;
   try {
@@ -78,15 +46,17 @@ function App(
 
   const [state, setState] = useSet({
     formData: {},
-    schema: {},
-    isNewVersion: true, // 用schema字段，还是用propsSchema字段，这是一个问题
-    selected: undefined, // 被选中的$id, 如果object/array的内部，以首字母0标识
+    frProps: {
+      displayType: 'row',
+    }, // form-render 的全局props等
     hovering: undefined, // 目前没有用到
+    isNewVersion: true, // 用schema字段，还是用propsSchema字段，这是一个问题
     preview: false, // preview = false 是编辑模式
-    ...initGlobal, // form-render 的全局props等
+    schema: {},
+    selected: undefined, // 被选中的$id, 如果object/array的内部，以首字母0标识
   });
 
-  // 在这里统一收口 propsSchema 到 schema 的转换
+  // 收口点 propsSchema 到 schema 的转换 (一共3处，其他两个是 importSchema 和 setValue，在 FRWrapper 文件)
   useEffect(() => {
     const schema = defaultValue ? transformFrom(defaultValue) : DEFAULT_SCHEMA;
     if (schema && schema.propsSchema) {
@@ -101,17 +71,18 @@ function App(
   }, [defaultValue]);
 
   const {
-    schema,
     formData,
-    preview,
-    selected,
+    frProps,
     hovering,
     isNewVersion,
-    ...rest
+    preview,
+    schema,
+    selected,
   } = state;
 
-  const { displayType } = rest;
+  const { displayType } = frProps;
   const showDescIcon = displayType === 'row' ? true : false;
+  const _frProps = { ...frProps, showDescIcon };
 
   const onChange = data => {
     setState({ formData: data });
@@ -125,33 +96,39 @@ function App(
 
   const _mapping = { ...mapping, array: 'listEditor' };
 
-  const globalProps = {
+  const rootState = {
     preview,
-    setState,
     simple: false,
     mapping: _mapping,
-    widgets,
+    widgets: { ...defaultWidgets, ...widgets },
     selected,
     hovering,
-    ...rest,
-    showDescIcon,
   };
 
-  const FRProps = {
-    schema,
-    formData,
-    onChange,
-    onSchemaChange,
+  const userProps = {
     templates,
     submit,
     transformFrom,
     transformTo,
     isNewVersion,
     extraButtons,
-    ...globalProps,
+    settings,
+    commonSettings,
+    globalSettings,
   };
 
-  return <FRWrapper ref={ref} {...FRProps} />;
+  const allProps = {
+    schema,
+    formData,
+    onChange,
+    setGlobal: setState,
+    onSchemaChange,
+    ...rootState, // 顶层的state
+    userProps, // 用户传入的props
+    frProps: _frProps, // fr顶层的props
+  };
+
+  return <FRWrapper ref={ref} {...allProps} />;
 }
 
 export default forwardRef(App);